requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2022.12.7
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
Django==5.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
frozenlist==1.6.0
geographiclib==2.0
geopy==2.3.0
graphene==3.4.1
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
kombu==5.4.2
multidict==6.4.3
mysql==0.0.3
mysql-connector==2.2.9
numpy==1.24.2
openpyxl==3.1.2
packaging==24.1
pandas==1.5.3
passlib==1.7.4
pgeocode==0.4.0
promise==2.3
prompt_toolkit==3.0.48
propcache==0.3.1
pycparser==2.22
pyfcm==1.5.4
PyJWT==2.9.0
PyMySQL==1.0.2
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.2
razorpay==1.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
twilio==9.6.1
typing_extensions==4.9.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
